% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_parameters}
\alias{global_parameters}
\title{Parameters Accessible through the Whole Book}
\format{
An object of class \code{list} of length 7.
}
\usage{
global_parameters
}
\description{
An R list of parameters that should be accessible throughout the entire book.
}
\keyword{datasets}
